Create/Edit/Delete actions (TaskStatuses|Labels) -> only for Auth users

+++ app/Http/Controllers/LabelController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Label;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class LabelController extends Controller
 {
@@ -19,8 +20,10 @@ class LabelController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        Gate::authorize('auth-for-crud', $request->getUser());
+
         $label = new Label();
         return view('label.create', compact('label'));
     }
@@ -30,6 +33,8 @@ class LabelController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('auth-for-crud', $request->getUser());
+
         $data = $this->validate($request, [
             'name' => 'required|min:1|unique:labels',
             'description' => 'nullable',
@@ -56,8 +61,10 @@ class LabelController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(Request $request, string $id)
     {
+        Gate::authorize('auth-for-crud', $request->getUser());
+
         $label = Label::findOrFail($id);
         return view('label.edit', compact('label'));
     }
@@ -67,6 +74,8 @@ class LabelController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        Gate::authorize('auth-for-crud', $request->getUser());
+
         $label = Label::findOrFail($id);
         $data = $this->validate($request, [
             'name' => 'required|min:1|unique:labels,name,' . $label->id,
@@ -85,8 +94,10 @@ class LabelController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
+        Gate::authorize('auth-for-crud', $request->getUser());
+
         $label = Label::find($id);
         if ($label) {
             $label->delete();
